Updated the delete feature of the pg, adding the responseentity and also adding the user not found exception

+++ src/main/java/com/rentpgapp/rent_pg_service/service/impl/PgServiceImpl.java
@@ -17,6 +17,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -58,15 +59,15 @@ public class PgServiceImpl implements PgService {
 //    }
 
     @Override
-    public boolean deletePgByNameAndLocation(String name, String location) {
-        try {
-            PayingGuestDetails pg = pgRepository.findByNameAndLocation(name, location)
-                    .orElseThrow(() -> new PgNotFoundException("PG not found with name: " + name + " location " + location ));
-            pgRepository.delete(pg);
-            return true; // successfully deleted
-        } catch (Exception e) {
-            throw new RuntimeException("Cannot delete PG with name " + name + " location " + location); // deletion failed
-        }
+    public boolean deletePgByNameAndLocation(Long ownerId, String name, String location) {
+        Users owner = userRepository.findById(ownerId)
+                .orElseThrow(() -> new UserNotFoundException("Owner with id " + ownerId + " not found!!"));
+        PayingGuestDetails pg = pgRepository.findByNameAndLocation(name, location)
+                .orElseThrow(() -> new PgNotFoundException("PG not found with name: " + name + " and location " + location + " for owner id " + ownerId));
+        if(!pg.getOwner().equals(owner))
+            return false;
+        pgRepository.delete(pg);
+        return true; // successfully deleted
     }
 
     @Override
